requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
pgvector
anthropic>=0.34.0
sentence-transformers
pymupdf
python-docx
markdown
beautifulsoup4
pytest
httpx
slowapi
rank-bm25
